test: move table expr coverage into test module
